requirements.txt
Flask==2.3.3
gunicorn==21.2.0
openai==1.3.0
python-dotenv==1.0.0
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
transformers==4.33.2
langchain==0.0.310
tiktoken==0.5.1
requests==2.31.0
pydantic==2.4.2
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
flask-cors==4.0.0 